// src/app/models/personalizeddetail.ts
import { Entrepreneurship } from './entrepreneurship';

export class PersonalizedDetail {
  idPersonalizedDetail: number = 0;
  namePersonalizedDetail: string = '';
  additionalPricePersonalizedDetail: number = 0;
  imagePersonalizedDetail: string = '';
  entrepreneurships: Entrepreneurship = new Entrepreneurship();
}